requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<23.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
whitenoise>=6.5,<7.0

# Testing
pytest>=7.4,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<6.0
factory-boy>=3.3,<4.0
responses>=0.24,<1.0
